+++ backend/src/main/java/com/colmena/belleza/repository/DetalleVentaRepository.java
@@ -3,4 +3,4 @@ package com.colmena.belleza.repository;
 import com.colmena.belleza.entity.DetalleVenta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Long> {}
+public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Integer> {}
