+++ backend/src/main/java/com/colmena/belleza/repository/ProveedorRepository.java
@@ -3,4 +3,4 @@ package com.colmena.belleza.repository;
 import com.colmena.belleza.entity.Proveedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProveedorRepository extends JpaRepository<Proveedor, Long> {}
+public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {}
