+++ backend/src/main/java/com/colmena/belleza/repository/VentaRepository.java
@@ -3,4 +3,4 @@ package com.colmena.belleza.repository;
 import com.colmena.belleza.entity.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VentaRepository extends JpaRepository<Venta, Long> {}
+public interface VentaRepository extends JpaRepository<Venta, Integer> {}
